Split NA Utils
