+++ app/src/main/java/ltd/nextalone/hook/HideOnlineStatus.kt
@@ -24,9 +24,9 @@ package ltd.nextalone.hook
 import android.widget.LinearLayout
 import android.widget.RelativeLayout
 import ltd.nextalone.util.clazz
-import ltd.nextalone.util.findHostViewById
+import ltd.nextalone.util.findHostView
+import ltd.nextalone.util.hide
 import ltd.nextalone.util.hookAfterAllConstructors
-import ltd.nextalone.util.setViewZeroSize
 import nil.nadph.qnotified.base.annotation.FunctionEntry
 import nil.nadph.qnotified.hook.CommonDelayableHook
 import nil.nadph.qnotified.util.Utils
@@ -38,7 +38,7 @@ object HideOnlineStatus : CommonDelayableHook("na_hide_online_status_kt") {
         return try {
             "com.tencent.mobileqq.widget.navbar.NavBarAIO".clazz.hookAfterAllConstructors {
                 val ctx = it.thisObject as RelativeLayout
-                ctx.findHostViewById<LinearLayout>("j65")!!.setViewZeroSize()
+                ctx.findHostView<LinearLayout>("j65")!!.hide()
             }
             true
         } catch (t: Throwable) {
