+++ app/src/main/java/me/nextalone/hook/DisabledRedNick.kt
@@ -38,7 +38,7 @@ object DisabledRedNick : CommonDelayableHook("na_disable_red_nick_kt") {
             if (!isSimpleUi) {
                 "Lcom/tencent/mobileqq/activity/aio/core/FriendChatPie;->aP()V".method.hookBefore(this) {
                     val navAIO = it.thisObject.get("a", "com.tencent.mobileqq.widget.navbar.NavBarAIO".clazz) as RelativeLayout
-                    val linearLayout = navAIO.findHostViewById<LinearLayout>("e89")
+                    val linearLayout = navAIO.findHostView<LinearLayout>("e89")
                     linearLayout?.hide()
                     it.result = null
                 }
